Update pagination
Add a small script that checks if the user is on the last page and add
a 'last-page' class to the 'nav-links' div. This will fix an issue when
the faux pagination button is visible on all pages when custom color is
set as a text color.

Update CSS to reflect the change described above and hide the faux
pgination buttom from all pages but last.